Enum annotate_snippets::display_list::DisplayMarkType

pub enum DisplayMarkType {
    AnnotationThrough,
    AnnotationStart,
}

A type of the DisplayMark.

Variants
AnnotationThrough

A mark indicating a multiline annotation going through the current line.

Example:

use annotate_snippets::display_list::*;
use annotate_snippets::formatter::DisplayListFormatter;

let dlf = DisplayListFormatter::new(false); // Don't use colors

let dl = DisplayList {
    body: vec![
        DisplayLine::Source {
            lineno: Some(51),
            inline_marks: vec![
                DisplayMark {
                    mark_type: DisplayMarkType::AnnotationThrough,
                    annotation_type: DisplayAnnotationType::Error,
                }
            ],
            line: DisplaySourceLine::Content {
                text: "Example".to_string(),
                range: (0, 7),
            }
        }
    ]
};
assert_eq!(dlf.format(&dl), "51 | | Example");

AnnotationStart

A mark indicating a multiline annotation starting on the given line.

Example:

use annotate_snippets::display_list::*;
use annotate_snippets::formatter::DisplayListFormatter;

let dlf = DisplayListFormatter::new(false); // Don't use colors

let dl = DisplayList {
    body: vec![
        DisplayLine::Source {
            lineno: Some(51),
            inline_marks: vec![
                DisplayMark {
                    mark_type: DisplayMarkType::AnnotationStart,
                    annotation_type: DisplayAnnotationType::Error,
                }
            ],
            line: DisplaySourceLine::Content {
                text: "Example".to_string(),
                range: (0, 7),
            }
        }
    ]
};
assert_eq!(dlf.format(&dl), "51 | / Example");