Enum annotate_snippets::display_list::DisplayRawLine

pub enum DisplayRawLine {
    Origin {
        path: String,
        pos: Option<(usize, usize)>,
        header_type: DisplayHeaderType,
    },
    Annotation {
        annotation: Annotation,
        source_aligned: bool,
        continuation: bool,
    },
}

Raw line - a line which does not have the lineno part and is not considered a part of the snippet.

Variants
Origin

A line which provides information about the location of the given slice in the project structure.

Fields of Origin

path: String

pos: Option<(usize, usize)>

header_type: DisplayHeaderType

Annotation

An annotation line which is not part of any snippet.

Fields of Annotation

annotation: Annotation

source_aligned: bool

If set to true, the annotation will be aligned to the lineno delimiter of the snippet.

continuation: bool

If set to true, only the label of the Annotation will be displayed. It allows for a multiline annotation to be aligned without displaing the meta information (type and id) to be displayed on each line.