annotate_snippets

Type Alias Report

pub type Report<'a> = &'a [Group<'a>];

A diagnostic message and any associated context to help users understand it

The first Group is the “primary” group, ie it contains the diagnostic message.

All subsequent Groups are for distinct pieces of context. The primary group will be visually distinguished to help tell them apart.