annotate_snippets

Struct Renderer

pub struct Renderer { /* private fields */ }

The Renderer for a Report

The caller is expected to detect any relevant terminal features and configure the renderer, including

ANSI Escape code support (always outputted with Renderer::styled)

Terminal width (Renderer::term_width)

Unicode support (Renderer::decor_style)

Example

let report = // ...

let renderer = Renderer::styled();
let output = renderer.render(report);
anstream::println!("{output}");

Implementations

impl Renderer

pub const fn plain() -> Self

No terminal styling

pub const fn styled() -> Self

Default terminal styling

If ANSI escape codes are not supported, either

Call Renderer::plain instead

Strip them after the fact, like with anstream

Note

When testing styled terminal output, see the testing-colors feature

pub const fn short_message(self, short_message: bool) -> Self

Abbreviate the message

pub const fn term_width(self, term_width: usize) -> Self

Set the width to render within

Affects the rendering of Snippets

pub const fn decor_style(self, decor_style: DecorStyle) -> Self

Set the character set used for rendering decor

pub const fn anonymized_line_numbers( self, anonymized_line_numbers: bool, ) -> Self

Anonymize line numbers

When enabled, line numbers are replaced with LL which is useful for tests.

Example

  --> $DIR/whitespace-trimming.rs:4:193
   |
LL | ...                   let _: () = 42;
   |                                   ^^ expected (), found integer
   |

impl Renderer

pub fn render(&self, groups: Report<'_>) -> String

Render a diagnostic Report

impl Renderer

Customize Renderer::styled

pub const fn error(self, style: Style) -> Self

Override the output style for error

pub const fn warning(self, style: Style) -> Self

Override the output style for warnings

pub const fn info(self, style: Style) -> Self

Override the output style for info

pub const fn note(self, style: Style) -> Self

Override the output style for notes

pub const fn help(self, style: Style) -> Self

Override the output style for help

pub const fn line_num(self, style: Style) -> Self

Override the output style for line numbers in the Snippet gutter

pub const fn emphasis(self, style: Style) -> Self

Override the output style for emphasis for the primary_title

pub const fn context(self, style: Style) -> Self

Override the output style for AnnotationKind::Context

pub const fn addition(self, style: Style) -> Self

Override the output style for Patch additions

pub const fn removal(self, style: Style) -> Self

Override the output style for Patch removals

pub const fn none(self, style: Style) -> Self

Override the output style for all other text